WIP assert Junit5 CustomEnvironmentSubstitutionCircularReferenceTest correct
